[refactor] 체험 상세 성능 개선
