+++ package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "analyze": "cross-env ANALYZE=true next build"
   },
   "dependencies": {
     "@tanstack/react-query": "^5.64.2",
@@ -27,10 +28,12 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
+    "@next/bundle-analyzer": "15.1.7",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "antd": "5.23.2",
+    "cross-env": "7.0.3",
     "eslint": "^9",
     "eslint-config-next": "15.1.4",
     "postcss": "^8",
